+++ R/ncUtils.R
@@ -160,8 +160,8 @@ dataIs180 <- function(data) {
 # just holds this dataframe so i can see it / add to it easily instead of storing it as an rdata
 getCoordNameMatch <- function() {
     data.frame(
-        current = c('lon', 'long', 'lat', 'time', 'longitude', 'latitude', 'utc', 'date', 'dayofyear', 'altitude', 'depth'),
-        standard = c('Longitude', 'Longitude', 'Latitude', 'UTC', 'Longitude', 'Latitude', 'UTC', 'UTC', 'UTC', 'Depth', 'Depth'),
+        current = c('lon', 'long', 'lat', 'time', 'longitude', 'latitude', 'utc', 'date', 'dayofyear', 'altitude', 'depth', 'level'),
+        standard = c('Longitude', 'Longitude', 'Latitude', 'UTC', 'Longitude', 'Latitude', 'UTC', 'UTC', 'UTC', 'Depth', 'Depth', 'Depth'),
         stringsAsFactors = FALSE
     )
 }
@@ -284,4 +284,10 @@ romsCheck <- function(nc) {
     # whichDimVar <- names(nc$var) %in% c('lon', 'lat', 'year', 'month', 'day')
     # nc$var <- nc$var[!whichDimVar]
     nc
+}
+
+nowUTC <- function() {
+    now <- Sys.time()
+    attr(now, 'tzone') <- 'UTC'
+    now
 }